Comments to your code
Amazing work. These are the things that you're doing great!

All the test cases! 
Good names in your variables. 
Clean code
The only thing that you should consider is that in some cases is a good idea to check if the input for the function is not valid. 
Keep doing this fantastic work!.